// src/types/shipper/transport.ts
export interface ITransportPlanSale {
  price: number;
  status: number;
  departure_from: number;
  arrival_to: number;
  transport_date: string;
  collection_time_from: string;
  collection_time_to: string;
  created_date: string;
  item_name_txt: string;
  trsp_cli_prty_name_txt: string;

  id: number;
  created_user: string;
  updated_user: string;
  updated_date: string;
  req_cns_line_item_id: number;
  operator_id: string;
  operator_code: string;
  trans_plan_id: string;
  transport_code: string;
  transport_name: string;
  operatorName: string;
  collection_date: string;
  trailer_number: number;
  trailer_number_rest: number;
  price_per_unit: number;
  vehicle_condition: number[];
  outer_package_code: number;
  total_length?: string;
  total_width?: string;
  total_height?: string;
  weight?: string;
  special_instructions?: string;
  trans_type: number;
  propose_price?: number;
  propose_depature_time?: string;
  propose_arrival_time?: string;
  istd_totl_vol_meas?: string;
  vol_unt_cd?: number;
  num_unt_cd?: number;
  weig_unt_cd?: number;
  trans_order_id?: number;
  cns_line_item_id?: number;
  shipper_operator_code?: string;
  temperatureRange: [];
}

export interface ITransportShipperSearch {
  id: number;
  day: string;
  price: number;
  status: number;
  giai: string;
  operator_id: string;
  operator_code: string;
  car_info_id: number;
  vehicle_availability_resource_id: number;
  departure_from: number;
  arrival_to: number;
  trip_name: string;
  departure_time_min: string;
  departure_time_max: string;
  arrival_time: string;
  adjustment_price?: number;
  vehicle_type?: number;
  display_order?: number;
  assign_type?: number;
  max_payload: number;
  total_length: number;
  total_width: number;
  total_height: number;
  vehicle_size?: number;
  service_strt_date: string;
  temperature_range?: [];
  cut_off_info_id?: number;
  cut_off_time?: number;
  cut_off_fee: number;
  free_time_info_id?: number;
  free_time_time?: number;
  free_time_fee?: number;
  trans_type?: number;
  operator_name: string;
  tractor_idcr: string;
  car_license_plt_num_id?: string;
  trailer_license_plt_num_id: string;
  service_no: string;
  created_date: string;
  vehicle_name: string;
  propose_trsp_plan: Proposetrspplan[];
}

interface Proposetrspplan {
  id: number;
  trans_type: number;
  operator_id: string;
  operator_name: string;
  cns_line_item_id: number;
  trans_plan_id: string;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  outer_package_code: number;
  price_per_unit: number;
  status: number;
  total_height: number;
  total_length: number;
  total_width: number;
  trailer_number: number;
  trailer_number_rest: number;
  weight: number;
  arrival_to: number;
  departure_from: number;
  transport_code: string;
  transport_name: string;
  special_instructions?: any;
  vehicle_condition?: any;
  created_date: string;
  shipper_info: Shipperinfo;
  trsp_plan_line_item: [];
}

interface Shipperinfo {
  id: number;
  operator_name: string;
  outer_package_code: string;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  special_instructions?: string;
}